backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg[binary]==3.2.12
resend==2.4.0
python-dotenv==1.0.1

